requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
tqdm>=4.65.0
numpy>=1.25.0
matplotlib>=3.7.0
ftfy>=6.1.1
regex>=2023.10.23
imagehash>=4.3.1
git+https://github.com/openai/CLIP.git